pyproject.toml
[project]
name = "crop-disease"
version = "0.1.0"
description = "Crop Disease Detection — Multi-Layer AI Pipeline (FastAPI)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "supabase>=2.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "openai-whisper",
    "sounddevice>=0.5.1",
    "numpy>=1.26.0",
    "google-genai>=0.3.0",
]

[project.scripts]
start = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
